Fix : function.ts 분리 및 버그수정
