+++ src/common/type.ts
@@ -1,3 +1,17 @@
 type ToolbarPostion = [number, number];
 
-export { ToolbarPostion };
+interface TextSegmentInfo {
+  tagInfos: TagInfo[];
+  innertext: string | null;
+}
+
+interface TagInfo {
+  name: string;
+  attributes: AttributeInfo[];
+}
+
+interface AttributeInfo {
+  name: string;
+  value: string;
+}
+export { ToolbarPostion, TextSegmentInfo, TagInfo, AttributeInfo };
